+++ src/main/kotlin/io/gitlab/arturbosch/detekt/idea/config/ui/DetektConfigUi.kt
@@ -5,8 +5,9 @@ import com.intellij.openapi.observable.properties.ObservableProperty
 import com.intellij.openapi.observable.util.bindEnabled
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogPanel
-import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.ui.dsl.builder.BottomGap
+import com.intellij.ui.dsl.builder.AlignX
+import com.intellij.ui.dsl.builder.AlignY
 import com.intellij.ui.dsl.builder.Cell
 import com.intellij.ui.dsl.builder.MutableProperty
 import com.intellij.ui.dsl.builder.Panel
@@ -14,13 +15,9 @@ import com.intellij.ui.dsl.builder.RowLayout
 import com.intellij.ui.dsl.builder.bindSelected
 import com.intellij.ui.dsl.builder.bindText
 import com.intellij.ui.dsl.builder.panel
-import com.intellij.ui.dsl.gridLayout.HorizontalAlign
-import com.intellij.ui.dsl.gridLayout.VerticalAlign
 import com.intellij.ui.layout.selected
 import io.gitlab.arturbosch.detekt.idea.DetektBundle
 import io.gitlab.arturbosch.detekt.idea.config.DetektPluginSettings
-import io.gitlab.arturbosch.detekt.idea.util.toPathsList
-import io.gitlab.arturbosch.detekt.idea.util.toVirtualFilesList
 import io.gitlab.arturbosch.detekt.idea.util.validateAsFilePath
 import java.io.File
 import javax.swing.JCheckBox
@@ -111,7 +108,7 @@ internal class DetektConfigUi(
     private fun Panel.configurationFilesRow(isEnabled: ObservableProperty<Boolean>) {
         row {
             val label = label(DetektBundle.message("detekt.configuration.configurationFiles.title"))
-                .verticalAlign(VerticalAlign.TOP)
+                .align(AlignY.TOP)
                 .component
 
             val listModel = FilesListPanel.ListModel()
@@ -126,7 +123,7 @@ internal class DetektConfigUi(
                 .bindEnabled(isEnabled)
 
             cell(filesListPanel)
-                .horizontalAlign(HorizontalAlign.FILL)
+                .align(AlignX.FILL)
                 .resizableColumn()
                 .bindItems(settings::configurationFilePaths, listModel)
 
@@ -141,21 +138,22 @@ internal class DetektConfigUi(
     private fun Panel.baselineFileRow() {
         row(DetektBundle.message("detekt.configuration.baselineFile.title")) {
             textFieldWithBrowseButton(
-                DetektBundle.message("detekt.configuration.baselineFile.dialog.title"),
-                project,
                 FileChooserDescriptorUtil.createSingleXmlChooserDescriptor()
+                    .withTitle(DetektBundle.message("detekt.configuration.baselineFile.dialog.title")),
+                project,
             )
                 .bindText(
-                    getter = { LocalFileSystem.getInstance().extractPresentableUrl(settings.baselinePath) },
+                    getter = { settings.baselinePath },
                     setter = {
-                        if (File(it).isFile) {
-                            settings.baselinePath = LocalFileSystem.getInstance().findFileByPath(it)?.path.orEmpty()
+                        val file = File(it)
+                        if (file.isFile) {
+                            settings.baselinePath = file.absolutePath
                         } else {
                             settings.baselinePath = ""
                         }
                     }
                 )
-                .horizontalAlign(HorizontalAlign.FILL)
+                .align(AlignX.FILL)
                 .resizableColumn()
                 .validationOnInput { validateAsFilePath(it.text, isWarning = true) }
                 .validationOnApply { validateAsFilePath(it.text) }
@@ -169,10 +167,10 @@ internal class DetektConfigUi(
     private fun Panel.pluginJarsRow(isEnabled: ObservableProperty<Boolean>) {
         row {
             val label = label(DetektBundle.message("detekt.configuration.pluginJarFiles.title"))
-                .verticalAlign(VerticalAlign.TOP)
+                .align(AlignY.TOP)
                 .component
 
-            val listModel = FilesListPanel.ListModel(settings.pluginJarPaths.toVirtualFilesList())
+            val listModel = FilesListPanel.ListModel(settings.pluginJarPaths)
             val filesListPanel = FilesListPanel(
                 listModel = listModel,
                 project = project,
@@ -184,7 +182,7 @@ internal class DetektConfigUi(
                 .bindEnabled(isEnabled)
 
             cell(filesListPanel)
-                .horizontalAlign(HorizontalAlign.FILL)
+                .align(AlignX.FILL)
                 .resizableColumn()
                 .bindItems(settings::pluginJarPaths, listModel)
 
@@ -202,13 +200,13 @@ internal class DetektConfigUi(
     ) {
         bind(
             { listModel.items },
-            { _, virtualFiles ->
+            { _, paths ->
                 listModel.clear()
-                listModel += virtualFiles
+                listModel += paths
             },
             MutableProperty(
-                { fileListProperty.get().toVirtualFilesList() },
-                { fileListProperty.set(it.toPathsList()) }
+                { fileListProperty.get() },
+                { fileListProperty.set(it) }
             )
         )
     }