+++ src/main/kotlin/io/gitlab/arturbosch/detekt/idea/problems/FindingsManager.kt
@@ -5,7 +5,7 @@ import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
-import io.gitlab.arturbosch.detekt.api.Finding
+import dev.detekt.api.Issue
 import java.util.concurrent.ConcurrentHashMap
 
 @Service(Service.Level.PROJECT)
@@ -17,7 +17,7 @@ class FindingsManager : Disposable {
     }
 
     private val listeners = ConcurrentHashMap.newKeySet<() -> Unit>()
-    private val state = ConcurrentHashMap<VirtualFile, List<Finding>>()
+    private val state = ConcurrentHashMap<VirtualFile, List<Issue>>()
 
     fun register(listener: () -> Unit) {
         listeners.add(listener)
@@ -29,11 +29,11 @@ class FindingsManager : Disposable {
 
     fun getAllFindingsSize(): Int = state.values.fold(0) { acc, cur -> acc + cur.size }
 
-    fun getFindings(file: VirtualFile): List<Finding> = state[file]?.toList() ?: emptyList()
+    fun getFindings(file: VirtualFile): List<Issue> = state[file]?.toList() ?: emptyList()
 
     fun getAnalyzedFiles(): Collection<VirtualFile> = state.keys.toList()
 
-    fun put(file: VirtualFile, findings: List<Finding>) {
+    fun put(file: VirtualFile, findings: List<Issue>) {
         state[file] = findings
     }